Rename Development.IDE.State to Development.IDE.Core
