+++ compiler/hie-core/hie-core.cabal
@@ -88,8 +88,8 @@ library
         Development.IDE.GHC.Error
         Development.IDE.Spans.Calculate
         Development.IDE.GHC.Warnings
-        Development.IDE.State.FileStore
-        Development.IDE.State.Rules
+        Development.IDE.Core.FileStore
+        Development.IDE.Core.Rules
         Development.IDE.GHC.Compat
         Development.IDE.LSP.LanguageServer
         Development.IDE.LSP.Definition
@@ -97,9 +97,9 @@ library
         Development.IDE.LSP.Protocol
         Development.IDE.LSP.Server
         Development.IDE.Types.Options
-        Development.IDE.State.RuleTypes
-        Development.IDE.State.Service
-        Development.IDE.State.Shake
+        Development.IDE.Core.RuleTypes
+        Development.IDE.Core.Service
+        Development.IDE.Core.Shake
         Development.IDE.Types.Diagnostics
         Development.IDE.Types.Location
         Development.IDE.Spans.Type
