+++ compiler/hie-core/src/Development/IDE/Core/Rules.hs
@@ -9,7 +9,7 @@
 -- | A Shake implementation of the compiler service, built
 --   using the "Shaker" abstraction layer for in-memory use.
 --
-module Development.IDE.State.Rules(
+module Development.IDE.Core.Rules(
     IdeState, GetDependencies(..), GetParsedModule(..), TransitiveDependencies(..),
     Priority(..),
     runAction, runActions, useE, usesE,
@@ -29,7 +29,7 @@ import qualified Development.IDE.Functions.Compile             as Compile
 import qualified Development.IDE.Types.Options as Compile
 import Development.IDE.Import.DependencyInformation
 import Development.IDE.Import.FindImports
-import           Development.IDE.State.FileStore
+import           Development.IDE.Core.FileStore
 import           Development.IDE.Types.Diagnostics as Base
 import Development.IDE.Types.Location
 import qualified Data.ByteString.UTF8 as BS
@@ -44,7 +44,7 @@ import qualified Data.Set                                 as Set
 import qualified Data.Text                                as T
 import           Development.IDE.GHC.Error
 import           Development.Shake                        hiding (Diagnostic, Env, newCache)
-import Development.IDE.State.RuleTypes
+import Development.IDE.Core.RuleTypes
 
 import           GHC
 import Development.IDE.GHC.Compat
@@ -52,8 +52,8 @@ import           UniqSupply
 import NameCache
 
 import qualified Development.IDE.Spans.AtPoint as AtPoint
-import Development.IDE.State.Service
-import Development.IDE.State.Shake
+import Development.IDE.Core.Service
+import Development.IDE.Core.Shake
 
 -- | This is useful for rules to convert rules that can only produce errors or
 -- a result into the more general IdeResult type that supports producing
