+++ compiler/hie-core/src/Development/IDE/LSP/LanguageServer.hs
@@ -18,7 +18,7 @@ import Control.Monad.IO.Class
 import qualified Development.IDE.LSP.Definition as LS.Definition
 import qualified Development.IDE.LSP.Hover      as LS.Hover
 import qualified Development.IDE.Types.Logger as Logger
-import Development.IDE.State.Service
+import Development.IDE.Core.Service
 import Development.IDE.Types.Location
 
 import qualified Data.Aeson                                as Aeson
@@ -26,7 +26,7 @@ import qualified Data.Rope.UTF16 as Rope
 import qualified Data.Set                                  as S
 import qualified Data.Text as T
 
-import Development.IDE.State.FileStore
+import Development.IDE.Core.FileStore
 
 import qualified Network.URI                               as URI
