+++ compiler/hie-core/test/Demo.hs
@@ -7,11 +7,11 @@ import Data.Maybe
 import Control.Concurrent.Extra
 import Control.Monad
 import System.Time.Extra
-import Development.IDE.State.FileStore
-import Development.IDE.State.Service
-import Development.IDE.State.Rules
-import Development.IDE.State.Shake
-import Development.IDE.State.RuleTypes
+import Development.IDE.Core.FileStore
+import Development.IDE.Core.Service
+import Development.IDE.Core.Rules
+import Development.IDE.Core.Shake
+import Development.IDE.Core.RuleTypes
 import Development.IDE.LSP.Protocol
 import Development.IDE.Types.Location
 import Data.String
