+++ daml-foundations/daml-ghc/ide/src/Development/IDE/State/Rules/Daml.hs
@@ -1,9 +1,9 @@
 -- Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 -- SPDX-License-Identifier: Apache-2.0
 {-# LANGUAGE OverloadedStrings #-}
-module Development.IDE.State.Rules.Daml
-    ( module Development.IDE.State.Rules
-    , module Development.IDE.State.Rules.Daml
+module Development.IDE.Core.Rules.Daml
+    ( module Development.IDE.Core.Rules
+    , module Development.IDE.Core.Rules.Daml
     ) where
 
 import Control.Concurrent.Extra
@@ -36,15 +36,15 @@ import qualified Network.HTTP.Types as HTTP.Types
 import qualified Network.URI as URI
 
 import Development.IDE.Import.DependencyInformation
-import Development.IDE.State.Rules hiding (mainRule)
-import qualified Development.IDE.State.Rules as IDE
-import Development.IDE.State.Service.Daml
-import Development.IDE.State.Shake
+import Development.IDE.Core.Rules hiding (mainRule)
+import qualified Development.IDE.Core.Rules as IDE
+import Development.IDE.Core.Service.Daml
+import Development.IDE.Core.Shake
 import Development.IDE.Types.Diagnostics
 import qualified Language.Haskell.LSP.Messages as LSP
 import qualified Language.Haskell.LSP.Types as LSP
 
-import Development.IDE.State.RuleTypes.Daml
+import Development.IDE.Core.RuleTypes.Daml
 
 import DA.Daml.GHC.Compiler.Convert (convertModule, sourceLocToRange)
 import DA.Daml.GHC.Compiler.UtilLF
